Правки задание 5
